# apis/v1/spec/swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: さくらのオブジェクトストレージ APIドキュメント
  contact:
    name: SAKURA internet Inc.
  license:
    name: Copyright(C) 2021 SAKURA internet Inc. all rights reserved.
  x-logo:
    url: logo.svg
    altText: さくらのオブジェクトストレージ：APIドキュメント
  description: |-

    ---

    「さくらのオブジェクトストレージ」が提供するAPIの利用方法とサンプルを公開しております。

    # 基本的な使い方

    ## APIキーの発行

    APIを利用するためには、認証のための「APIキー」が必要です。事前にキーを発行しておきます。
    APIキーは「ユーザーID」「パスワード」に相当する「トークン」と呼ばれる認証情報で構成されています。

    |   項目名   | APIキー発行時の項目名        | このドキュメント内での例             |
    |------------|------------------------------|--------------------------------------|
    | ユーザーID | アクセストークン(UUID)       | 01234567-89ab-cdef-0123-456789abcdef |
    | パスワード | アクセストークンシークレット | SAMPLETOKENSAMPLETOKENSAMPLETOKENSAM |

    <div class="warning">
    <b>操作マニュアル</b><br />
    <ul><li><a href="https://manual.sakura.ad.jp/cloud/api/apikey.html">APIキー | さくらのクラウド ドキュメント</a></li></ul>
    </div>

    ## 入力パラメータ

    APIの入力には送信先URLに対して、いくつかのヘッダーとAPIキーを送信します。

    * APIのURLは以下の2つが存在します。※ 各APIの使い分けは後述します。
      * `https://secure.sakura.ad.jp/cloud/zone/is1a/api/objectstorage/1.0/fed/v1/(エンドポイント)`
      * `https://secure.sakura.ad.jp/cloud/zone/is1a/api/objectstorage/1.0/（サイト名）/v2/(エンドポイント)`
    * 認証方式はHTTP Basic認証です。APIキーのアクセストークンをユーザーID、アクセストークンシークレットをパスワードとして指定します。

    ```
    # 入力サンプル
    curl -u '01234567-89ab-cdef-0123-456789abcdef:SAMPLETOKENSAMPLETOKENSAMPLETOKENSAM' \
...
    ```

    ## 出力結果と応答コード（HTTPステータスコード）

...

...

...

    ```
...

...

...
paths:
  /fed/v1/clusters:
    get:
      operationId: getClusters
      tags:
        - サイト
      summary: サイト一覧の取得
      description: |-
        アクセス先ベースURL: `https://secure.sakura.ad.jp/cloud/zone/is1a/api/objectstorage/1.0/fed/v1/`

        上記アクセス先ベースURLと右ペインのパスを組み合わせたものがアクセス先URLになります。
        例えば、右ペインのパスが `/clusters` の場合、アクセス先URLは、
        `https://secure.sakura.ad.jp/cloud/zone/is1a/api/objectstorage/1.0/fed/v1/clusters`
        となります。
      responses:
        '200':
          description: サイト一覧の取得に成功しました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListClustersResponseBody'
        '401':
          description: 認証に失敗しました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error401'
  '/fed/v1/clusters/{site_id}':
    get:
      operationId: getCluster
      tags:
        - サイト
      summary: サイトの取得
      description: |-
        アクセス先ベースURL: `https://secure.sakura.ad.jp/cloud/zone/is1a/api/objectstorage/1.0/fed/v1/`

        上記アクセス先ベースURLと右ペインのパスを組み合わせたものがアクセス先URLになります。
        例えば、右ペインのパスが `/clusters` の場合、アクセス先URLは、
        `https://secure.sakura.ad.jp/cloud/zone/is1a/api/objectstorage/1.0/fed/v1/clusters`
        となります。
      parameters:
        - in: path
          name: site_id
          required: true
          schema:
            type: string
          description: サイト名
      responses:
        '200':
          description: サイトの取得に成功しました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadClusterResponseBody'
        '401':
          description: 認証に失敗しました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error401'
        '404':
          description: サイトが見つかりませんでした
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
  '/fed/v1/buckets/{bucket_name}':
    put:
      operationId: createBucket
      tags:
        - バケット
      summary: バケットの作成
      description: |-
        アクセス先ベースURL: `https://secure.sakura.ad.jp/cloud/zone/is1a/api/objectstorage/1.0/fed/v1/`

        上記アクセス先ベースURLと右ペインのパスを組み合わせたものがアクセス先URLになります。
        例えば、右ペインのパスが `/clusters` の場合、アクセス先URLは、
        `https://secure.sakura.ad.jp/cloud/zone/is1a/api/objectstorage/1.0/fed/v1/clusters`
        となります。
      parameters:
        - in: path
          name: bucket_name
          required: true
          schema:
            $ref: '#/components/schemas/BucketName'
          description: バケット名
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBucketRequestBody'
        description: 作成したいバケット情報
        required: true
      responses:
        '201':
          description: バケット作成に成功しました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateBucketResponseBody'
        '400':
          description: バケット名が不正です
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
        '404':
          description: サイトアカウントが見つかりませんでした
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
        '409':
          description: 同名バケットがすでに存在します
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error409'
    delete:
      operationId: deleteBucket
      tags:
        - バケット
      servers:
        - url: 'https://secure.sakura.ad.jp/cloud/zone/is1a/api/objectstorage/1.0/fed/v1/'
      summary: バケットの削除
      description: |-
        アクセス先ベースURL: `https://secure.sakura.ad.jp/cloud/zone/is1a/api/objectstorage/1.0/fed/v1/`

        上記アクセス先ベースURLと右ペインのパスを組み合わせたものがアクセス先URLになります。
        例えば、右ペインのパスが `/clusters` の場合、アクセス先URLは、
        `https://secure.sakura.ad.jp/cloud/zone/is1a/api/objectstorage/1.0/fed/v1/clusters`
        となります。
      parameters:
        - in: path
          name: bucket_name
          required: true
          schema:
            $ref: '#/components/schemas/BucketName'
          description: バケット名
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBucketRequestBody'
        description: 作成したいバケット情報
        required: true
      responses:
        '204':
          description: バケット削除に成功しました
        '400':
          description: バケット名が不正です
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400'
        '409':
          description: バケットに関係するリソースが残っています、先にそれらのリソースを削除する必要があります
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error409'
  '/{site_id}/v2/account':
    get:
      operationId: getAccount
      tags:
        - サイトアカウント
      summary: サイトアカウントの取得
      description: |-
        アクセス先ベースURL: `https://secure.sakura.ad.jp/cloud/zone/is1a/api/objectstorage/1.0/（サイト名）/v2/`

        上記アクセス先ベースURLと右ペインのパスを組み合わせたものがアクセス先URLになります。
        例えば、右ペインのパスが `/account` の場合、アクセス先URLは、
        `https://secure.sakura.ad.jp/cloud/zone/is1a/api/objectstorage/1.0/（サイト名）/v2/account`
        となります。

        なお、サイト名の取得方法は「利用例＞接続先サイト一覧の取得」をご参考下さい。
      parameters:
        - in: path
          name: site_id
          required: true
          schema:
            type: string
          description: サイト名
      responses:
        '200':
          description: サイトアカウントの取得に成功しました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponseBody'
        '401':
          description: 認証に失敗しました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error401'
        '404':
          description: サイトアカウントが見つかりませんでした
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
        default:
          description: 不明なエラーです
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDefault'
    post:
      operationId: createAccount
      tags:
        - サイトアカウント
      summary: サイトアカウントの作成
      description: |-
        アクセス先ベースURL: `https://secure.sakura.ad.jp/cloud/zone/is1a/api/objectstorage/1.0/（サイト名）/v2/`

        上記アクセス先ベースURLと右ペインのパスを組み合わせたものがアクセス先URLになります。
        例えば、右ペインのパスが `/account` の場合、アクセス先URLは、
        `https://secure.sakura.ad.jp/cloud/zone/is1a/api/objectstorage/1.0/（サイト名）/v2/account`
        となります。

        なお、サイト名の取得方法は「利用例＞接続先サイト一覧の取得」をご参考下さい。
      parameters:
        - in: path
          name: site_id
          required: true
          schema:
            type: string
          description: サイト名
      responses:
        '201':
          description: サイトアカウントの作成に成功しました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponseBody'
        '401':
          description: 認証に失敗しました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error401'
        '403':
          description: 新規のサイトアカウント作成が拒否されました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error403'
        '409':
          description: サイトアカウントはすでに作成済みです
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error409'
        default:
          description: 不明なエラーです
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDefault'
    delete:
      operationId: deleteAccount
      tags:
        - サイトアカウント
      summary: サイトアカウントの削除
      description: |-
        アクセス先ベースURL: `https://secure.sakura.ad.jp/cloud/zone/is1a/api/objectstorage/1.0/（サイト名）/v2/`

        上記アクセス先ベースURLと右ペインのパスを組み合わせたものがアクセス先URLになります。
        例えば、右ペインのパスが `/account` の場合、アクセス先URLは、
        `https://secure.sakura.ad.jp/cloud/zone/is1a/api/objectstorage/1.0/（サイト名）/v2/account`
        となります。

        なお、サイト名の取得方法は「利用例＞接続先サイト一覧の取得」をご参考下さい。
      parameters:
        - in: path
          name: site_id
          required: true
          schema:
            type: string
          description: サイト名
      responses:
        '204':
          description: サイトアカウントの削除に成功しました
        '401':
          description: 認証に失敗しました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error401'
        '409':
          description: サイトアカウントに関係するリソースが残っています、先にそれらのリソースを削除する必要があります
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error409'
...
  '/{site_id}/v2/account/keys/{account_key_id}':
    get:
      operationId: getAccountKey
      tags:
        - サイトアカウント
      summary: サイトアカウントのアクセスキーの取得
      description: |-
        アクセス先ベースURL: `https://secure.sakura.ad.jp/cloud/zone/is1a/api/objectstorage/1.0/（サイト名）/v2/`

        上記アクセス先ベースURLと右ペインのパスを組み合わせたものがアクセス先URLになります。
        例えば、右ペインのパスが `/account` の場合、アクセス先URLは、
        `https://secure.sakura.ad.jp/cloud/zone/is1a/api/objectstorage/1.0/（サイト名）/v2/account`
        となります。

        なお、サイト名の取得方法は「利用例＞接続先サイト一覧の取得」をご参考下さい。
      parameters:
        - in: path
          name: site_id
          required: true
          schema:
            type: string
          description: サイト名
        - in: path
          name: account_key_id
          schema:
            $ref: '#/components/schemas/AccessKeyID'
          required: true
          description: アクセスキーID
      responses:
        '200':
          description: サイトアカウントのアクセスキーの取得に成功しました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountKeyResponseBody'
        '401':
          description: 認証に失敗しました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error401'
        '404':
          description: アクセスキーが見つかりませんでした
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404'
        default:
          description: 不明なエラーです
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDefault'
    delete:
      operationId: deleteAccountKey
      tags:
        - サイトアカウント
      summary: サイトアカウントのアクセスキーの削除
      description: |-
        アクセス先ベースURL: `https://secure.sakura.ad.jp/cloud/zone/is1a/api/objectstorage/1.0/（サイト名）/v2/`

        上記アクセス先ベースURLと右ペインのパスを組み合わせたものがアクセス先URLになります。
        例えば、右ペインのパスが `/account` の場合、アクセス先URLは、
        `https://secure.sakura.ad.jp/cloud/zone/is1a/api/objectstorage/1.0/（サイト名）/v2/account`
        となります。

        なお、サイト名の取得方法は「利用例＞接続先サイト一覧の取得」をご参考下さい。
      parameters:
        - in: path
          name: site_id
          required: true
          schema:
            type: string
          description: サイト名
        - in: path
          name: account_key_id
          schema:
            $ref: '#/components/schemas/AccessKeyID'
          required: true
          description: アクセスキーID
      responses:
        '204':
          description: サイトアカウントのアクセスキーの削除に成功しました
        '401':
          description: 認証に失敗しました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error401'
...
servers:
...
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
      description: |-
        [クラウドコントロールパネル](https://secure.sakura.ad.jp/cloud/)で発行したAPIキーを利用します。

         `アクセストークン(UUID)` を「ユーザーID」、`アクセストークンシークレット` を「パスワード」として指定します。
  schemas:
...
    ReadClusterResponseBody:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Cluster'
    # Note: handler.listClustersResからリネーム
    ListClustersResponseBody:
      type: object
      properties:
        data:
          description: 'If use a pointer type, braek output'
          type: array
          items:
            $ref: '#/components/schemas/Cluster'
      required:
        - data
    # Note: handler.putBucketRes:
...
    Error401:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
      required:
        - error
...
    Error404:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
      required:
        - error
...